requirements.txt
azure-identity==1.6.0
streamlit==1.18.1
openai==0.27.1
python-dotenv==0.21.0
numpy
pandas
matplotlib==3.6.3
plotly==5.12.0
scipy==1.10.0
scikit-learn==1.2.0
tenacity
tiktoken==0.3.0
llama-index==0.4.33
langchain==0.0.129
